Add an expiration tag on S3 dumps
So instead of adding a global expiration on the s3 bucket, we can
configure a rule to delete the expired objects having this tag.
It allows us to push objects that we want to persist as soon as we don't
put the tag (or we can remove the tag from a particular object).